smoothing lines
