+++ energy_demand/charts/resilience_multiple_lines.py
@@ -5,8 +5,7 @@ import os
 from collections import defaultdict
 import matplotlib.pyplot as plt
 from energy_demand.plotting import plotting_styles
-from energy_demand.plotting import plotting_results
-from energy_demand.plotting import plotting_program
+from energy_demand.plotting import basic_plot_functions
 from energy_demand.read_write import write_data
 from energy_demand.basic import basic_functions
 
@@ -47,6 +46,8 @@ def read_csv_average(path_to_csv):
     return out_dict
 
 def generate_min_max_resilience_plot(path_to_scenarios):
+    """
+    """
     # Iterate result folder
 
 
@@ -54,13 +55,16 @@ def generate_min_max_resilience_plot(path_to_scenarios):
     scenarios = os.listdir(path_to_scenarios)
 
 
-    # NOTE: the average scenario needs to be labelled 'av' and must not be at first position
+    # The average scenario needs to be labelled 'av' and must not be at first position
     if scenarios[0] == 'av':
         del scenarios[0]
         scenarios.append('av')
 
     data_container = {}
-    statistics_to_print = ["scenario \tmin \tmax \tdiff \tDate maximum day \tDate minimum day \tMax hour \tMin hour", "=============================", " "]
+    statistics_to_print = [
+        "scenario \tmin \tmax \tdiff \tDate maximum day \tDate minimum day \tMax hour \tMin hour",
+        "=============================",
+        " "]
 
     for scenario in scenarios:
         if scenario == '__results_resilience':
@@ -150,6 +154,7 @@ def generate_min_max_resilience_plot(path_to_scenarios):
             peak_day_values_to_print.append("======================")
             peak_day_values_to_print.append(str(data_container[scenario][2015]['max_values']))
             cnt += 1
+
     # ------------------
     # Write statiscs to txt
     # --------------------
@@ -163,9 +168,6 @@ def generate_min_max_resilience_plot(path_to_scenarios):
         os.path.join(path_to_scenarios, "__results_resilience", "values_peak_day_scenario.txt"),
         peak_day_values_to_print)
 
-    # Average demand of base year
-    flat_y_data = 25 * [average_demands_by]
-
     # ------------------------------------------------
     # Create maximum plot
     #
@@ -175,13 +177,28 @@ def generate_min_max_resilience_plot(path_to_scenarios):
     # ------------------------------------------------
     ymax = 60
     line_width = 1.0
-    color_flat_line = 'blue' #darkkhaki'
-    x_data = range(25)
+    #x_data = range(25)
+    x_data = range(24)
 
     fig_name = os.path.join(path_to_scenarios, "__results_resilience", "min_max_days.pdf")
 
     colors = plotting_styles.color_list_resilience()
-    fig = plt.figure(figsize=plotting_program.cm2inch(8, 8)) #width, height
+
+    colors = [
+        '#d7191c',
+        '#2c7bb6',
+        '#fdae61',
+        '#abd9e9']
+
+    '''colors = [
+        '#a6611a',
+        '#dfc27d',
+        '#80cdc1',
+        '#018571'
+        ]'''
+
+
+    fig = plt.figure(figsize=basic_plot_functions.cm2inch(8, 8)) #width, height
 
     for counter, (scenario, scenario_data) in enumerate(data_container.items()):
 
@@ -192,9 +209,10 @@ def generate_min_max_resilience_plot(path_to_scenarios):
         if counter == 0:
 
             # Plot current year maximum day plot
-            day_24h = [scenario_data[2015]['max_values'][-1]] + scenario_data[2015]['max_values']
+            #day_24h = [scenario_data[2015]['max_values'][-1]] + scenario_data[2015]['max_values']
+            day_24h = scenario_data[2015]['max_values']
 
-            x_data_smoothed, y_data_smoothed = plotting_results.smooth_data(x_data, day_24h, num=40000)
+            x_data_smoothed, y_data_smoothed = basic_plot_functions.smooth_data(x_data, day_24h, num=40000)
             plt.plot(
                 x_data_smoothed,
                 list(y_data_smoothed),
@@ -204,8 +222,9 @@ def generate_min_max_resilience_plot(path_to_scenarios):
                 label='{}__{}'.format(scenario, plot_yr))
 
             # Plot current year minimum day plot 
-            day_24h = [scenario_data[2015]['min_values'][-1]] + scenario_data[2015]['min_values']
-            x_data_smoothed, y_data_smoothed = plotting_results.smooth_data(x_data, day_24h, num=40000)
+            #day_24h = [scenario_data[2015]['min_values'][-1]] + scenario_data[2015]['min_values']
+            day_24h = scenario_data[2015]['min_values']
+            x_data_smoothed, y_data_smoothed = basic_plot_functions.smooth_data(x_data, day_24h, num=40000)
 
             plt.plot(
                 x_data_smoothed,
@@ -215,21 +234,13 @@ def generate_min_max_resilience_plot(path_to_scenarios):
                 linewidth=1.5,
                 label='{}__{}'.format(scenario, plot_yr))
 
-            # Add flat line
-            plt.plot(
-                x_data,
-                list(flat_y_data),
-                color=color_flat_line,
-                linestyle='--',
-                linewidth=line_width,
-                label='{}__{}'.format("flat", 2015))
-            
         y_data_max = scenario_data[plot_yr]['max_values']
         y_data_min = scenario_data[plot_yr]['min_values']
 
         # Maximum day line
-        day_24h = [y_data_max[-1]] + y_data_max
-        x_data_smoothed, y_data_smoothed = plotting_results.smooth_data(x_data, day_24h, num=40000)
+        #day_24h = [y_data_max[-1]] + y_data_max
+        day_24h = y_data_max
+        x_data_smoothed, y_data_smoothed = basic_plot_functions.smooth_data(x_data, day_24h, num=40000)
 
         plt.plot(
             x_data_smoothed,
@@ -240,8 +251,9 @@ def generate_min_max_resilience_plot(path_to_scenarios):
             label='{}__{}'.format(scenario, plot_yr))
 
         # Minimum day line
-        day_24h = [y_data_min[-1]] + y_data_min
-        x_data_smoothed, y_data_smoothed = plotting_results.smooth_data(x_data, day_24h, num=40000)
+        #day_24h = [y_data_min[-1]] + y_data_min
+        day_24h = y_data_min
+        x_data_smoothed, y_data_smoothed = basic_plot_functions.smooth_data(x_data, day_24h, num=40000)
 
         plt.plot(
             x_data_smoothed,
@@ -256,8 +268,12 @@ def generate_min_max_resilience_plot(path_to_scenarios):
     plt.xlim(xmin=0, xmax=23)
 
     # 24 hour day ticks
-    plt.xticks([0,6,12,18,24], [0,6,12,18,24])
+    #plt.xticks([0,6,12,18,24], [0,6,12,18,24])
+    #plt.xticks([0,5,11,17,23], [0,6,12,18,24])
 
+    # Position  Label
+    #plt.xticks([0,2,5,11,17,19, 23], [1,3,6,12,18,20,24])
+    plt.xticks([0,5,11,17,23], [1,6,12,18,24])
     # Legend #https://stackoverflow.com/questions/4700614/how-to-put-the-legend-out-of-the-plot
     ax = plt.subplot(111)
     box = ax.get_position()
@@ -283,7 +299,7 @@ def generate_min_max_resilience_plot(path_to_scenarios):
 
     colors = plotting_styles.color_list_resilience()
     fig = plt.figure(
-        figsize=plotting_program.cm2inch(8, 8)) #width, height
+        figsize=basic_plot_functions.cm2inch(8, 8)) #width, height
 
     for counter, (scenario, scenario_data) in enumerate(data_container.items()):
         x_data = range(24)
@@ -293,7 +309,7 @@ def generate_min_max_resilience_plot(path_to_scenarios):
     
         # Plot base year line
         if counter == 0:
-            x_data_smoothed, y_data_smoothed = plotting_results.smooth_data(x_data, scenario_data[2015]['min_values'], num=40000)
+            x_data_smoothed, y_data_smoothed = basic_plot_functions.smooth_data(x_data, scenario_data[2015]['min_values'], num=40000)
             plt.plot(
                 x_data_smoothed,
                 list(y_data_smoothed),
@@ -314,7 +330,7 @@ def generate_min_max_resilience_plot(path_to_scenarios):
         y_data = scenario_data[plot_yr]['min_values']
 
         # smooth line
-        x_data_smoothed, y_data_smoothed = plotting_results.smooth_data(x_data, y_data, num=40000)
+        x_data_smoothed, y_data_smoothed = basic_plot_functions.smooth_data(x_data, y_data, num=40000)
         plt.plot(
             x_data_smoothed,
             list(y_data_smoothed),
