+++ energy_demand/config_data/_scrap/dm_improvement.csv
@@ -1,0 +1,3 @@
+sub_param_name,default_by,end_yr,value_ey
+rs_lighting,0,2020,0.3
+rs_lighting,0,2050,0.5
