+++ energy_demand/config_data/_scrap/switches_fuel.csv
