+++ energy_demand/geography/spatial_diffusion.py
@@ -51,8 +51,8 @@ def spatial_diffusion_values(
         # ----------------
         # plot real values to check for outliers
         # ----------------
-        #from energy_demand.plotting import plotting_program
-        #plotting_program.plot_xy(list(real_values.values()))
+        #from energy_demand.plotting import basic_plot_functions
+        #basic_plot_functions.plot_xy(list(real_values.values()))
 
         # Select number of outliers to remove lower and higher extremes
         nr_of_outliers = int(100 / len(regions) * p_outlier)
@@ -71,7 +71,7 @@ def spatial_diffusion_values(
                 real_values[reg] = treshold_lower_real_value
 
         # Plot after removing outliers
-        #plotting_program.plot_xy(list(real_values.values()))
+        #basic_plot_functions.plot_xy(list(real_values.values()))
 
         # ---------------------------------
         # Congruence calculations
